MYR-320 Extract snapshotRowFromVehicle so the /snapshot projection is testable
The store.Vehicle -> telemetry.VehicleSnapshotRow projection was an inline
struct literal inside vehicleSnapshotAdapter.GetByID, reachable only through a
*store.VehicleRepo and therefore only with a live database. That is why it is
the one layer of the /snapshot read path with no unit test.

Pure move, no behaviour change: the literal is lifted verbatim into a free
function in its own file (adapters.go was 605 lines, well past the 300-line
cap; it is now 512 and the new file is 120).
