+++ cmd/telemetry-server/adapters.go
@@ -220,100 +220,6 @@ func (a *vehicleListerAdapter) ListByUser(ctx context.Context, userID string) ([
 	return out, nil
 }
 
-// vehicleSnapshotAdapter adapts store.VehicleRepo.GetByID to the
-// telemetry.VehicleSnapshotReader interface used by the
-// GET /api/vehicles/{vehicleId}/snapshot and
-// GET /api/vehicles/{vehicleId}/drives handlers (MYR-133). The
-// adapter exists so the handlers stay decoupled from internal/store.
-//
-// GetByID returns the wide read with GPS dual-resolution and
-// nav-route blob decryption already applied per
-// vehicle_repo_scan.go — exactly what /snapshot needs. The drives
-// handler reuses this for the ownership check (it needs the userID
-// off the row; GetByID is the cheapest source of truth for an
-// arbitrary vehicleId cuid).
-type vehicleSnapshotAdapter struct {
-	repo *store.VehicleRepo
-}
-
-func (a *vehicleSnapshotAdapter) GetByID(ctx context.Context, vehicleID string) (telemetry.VehicleSnapshotRow, error) {
-	v, err := a.repo.GetByID(ctx, vehicleID)
-	if err != nil {
-		return telemetry.VehicleSnapshotRow{}, fmt.Errorf("get vehicle by id: %w", err)
-	}
-	return telemetry.VehicleSnapshotRow{
-		ID:                   v.ID,
-		UserID:               v.UserID,
-		VIN:                  v.VIN,
-		Name:                 v.Name,
-		Model:                v.Model,
-		Year:                 v.Year,
-		Color:                v.Color,
-		LicensePlate:         v.LicensePlate,
-		Status:               string(v.Status),
-		ChargeLevel:          v.ChargeLevel,
-		EstimatedRange:       v.EstimatedRange,
-		ChargeState:          v.ChargeState,
-		TimeToFull:           v.TimeToFull,
-		Speed:                v.Speed,
-		GearPosition:         v.GearPosition,
-		Heading:              v.Heading,
-		Latitude:             v.Latitude,
-		Longitude:            v.Longitude,
-		LocationName:         v.LocationName,
-		LocationAddress:      v.LocationAddress,
-		InteriorTemp:         v.InteriorTemp,
-		ExteriorTemp:         v.ExteriorTemp,
-		OdometerMiles:        v.OdometerMiles,
-		FsdMilesSinceReset:   v.FsdMilesSinceReset,
-		DestinationName:      v.DestinationName,
-		DestinationAddress:   v.DestinationAddress,
-		DestinationLatitude:  v.DestinationLatitude,
-		DestinationLongitude: v.DestinationLongitude,
-		OriginLatitude:       v.OriginLatitude,
-		OriginLongitude:      v.OriginLongitude,
-		EtaMinutes:           v.EtaMinutes,
-		TripDistRemaining:    v.TripDistRemaining,
-		NavRouteCoordinates:  v.NavRouteCoordinates,
-		LastUpdated:          v.LastUpdated,
-		Locked:               v.IsLocked,
-		FrunkOpen:            v.FrunkOpen,
-		TrunkOpen:            v.TrunkOpen,
-		IsClimateOn:          v.IsClimateOn,
-		ChargePortDoorOpen:   v.ChargePortOpen,
-		DriverTempSetting:    v.DriverTempSetting,
-		PassengerTempSetting: v.PassengerTempSetting,
-		FanSpeed:             v.FanSpeed,
-		SeatHeaterLeft:       v.SeatHeaterLeft,
-		SeatHeaterRight:      v.SeatHeaterRight,
-		SeatHeaterRearLeft:   v.SeatHeaterRearLeft,
-		SeatHeaterRearCenter: v.SeatHeaterRearCenter,
-		SeatHeaterRearRight:  v.SeatHeaterRearRight,
-		SeatCoolerLeft:       v.SeatCoolerLeft,
-		SeatCoolerRight:      v.SeatCoolerRight,
-		MediaVolume:          v.MediaVolume,
-		SoftwareVersion:      v.SoftwareVersion,
-		Trim:                 v.Trim,
-		HvacAutoMode:         v.HvacAutoMode,
-		HvacAcEnabled:        v.HvacAcEnabled,
-		SeatVentEnabled:      v.SeatVentEnabled,
-		MediaPlaybackStatus:  v.MediaPlaybackStatus,
-
-		// MYR-303 media now-playing + MYR-308 seat-cooling capability.
-		MediaNowPlayingTitle:    v.MediaNowPlayingTitle,
-		MediaNowPlayingArtist:   v.MediaNowPlayingArtist,
-		MediaNowPlayingAlbum:    v.MediaNowPlayingAlbum,
-		MediaNowPlayingStation:  v.MediaNowPlayingStation,
-		MediaPlaybackSource:     v.MediaPlaybackSource,
-		MediaNowPlayingDuration: v.MediaNowPlayingDuration,
-		MediaNowPlayingElapsed:  v.MediaNowPlayingElapsed,
-		MediaVolumeMax:          v.MediaVolumeMax,
-		SeatCoolingCapable:      v.SeatCoolingCapable,
-		ServiceETC:              v.ServiceETC,
-		ServiceExpectedEndAt:    v.ServiceExpectedEndAt,
-	}, nil
-}
-
 // driveListerAdapter adapts store.DriveRepo.ListByVehicleID to the
 // telemetry.DriveLister interface used by the drives handler
 // (MYR-133). The handler defines its own DriveListCursor /
